2Week_Mission/src/test/java/com/ll/exam/final__2022_10_08/service/OrderServiceTests.java
package com.ll.exam.final__2022_10_08.service;

import com.ll.exam.final__2022_10_08.app.cart.entity.CartItem;
import com.ll.exam.final__2022_10_08.app.cart.service.CartService;
import com.ll.exam.final__2022_10_08.app.member.entity.Member;
import com.ll.exam.final__2022_10_08.app.member.service.MemberService;
import com.ll.exam.final__2022_10_08.app.order.entity.Order;
import com.ll.exam.final__2022_10_08.app.order.entity.OrderService;
import java.util.List;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.test.context.ActiveProfiles;
import org.springframework.transaction.annotation.Transactional;

import static org.assertj.core.api.Assertions.assertThat;

@SpringBootTest
@Transactional
@ActiveProfiles("test")
public class OrderServiceTests {
    @Autowired
    private MemberService memberService;
    @Autowired
    private OrderService orderService;
    @Autowired
    private CartService cartService;
    @Test
    @DisplayName("주문 생성")
    @Rollback(value = false)
    public void t1() {
        // Given
        Member member = memberService.findByUsername("user2").get();
        List<CartItem> cartItems = cartService.findAllByMemberId(member.getId());

        // When
        Order order = orderService.createFromCart(member);

        // Then
        assertThat(order.getOrderItems().size()).isEqualTo(cartItems.size());
    }
}
